rpaleoclim: Download Paleoclimate Data from 'PaleoClim'

'PaleoClim' http://www.paleoclim.org (Brown et al. 2019, doi:10.1038/sdata.2018.254 ) is a set of free, high resolution paleoclimate surfaces covering the whole globe. It includes data on surface temperature, precipitation and the standard bioclimatic variables commonly used in ecological modelling, derived from the 'HadCM3' general circulation model and downscaled to a spatial resolution of up to 2.5 minutes. Simulations are available for key time periods from the Late Holocene to mid-Pliocene. Data on current and Last Glacial Maximum climate is derived from 'CHELSA' (Karger et al. 2017, doi:10.1038/sdata.2017.122 ) and reprocessed by 'PaleoClim' to match their format; it is available at up to 30 seconds resolution. This package provides a simple interface for downloading 'PaleoClim' data in R, with support for caching and filtering retrieved data by period, resolution, and geographic extent.

See also

Author

Maintainer: Joe Roe joe@joeroe.io [copyright holder]